refactor: move deployment-loader types to internal
