+++ packages/core/src/new-api/internal/deployer.ts
@@ -17,7 +17,6 @@ import {
 } from "../type-guards";
 import { Artifact, ArtifactResolver } from "../types/artifact";
 import { DeployConfig, DeploymentResult } from "../types/deployer";
-import { DeploymentLoader } from "../types/deployment-loader";
 
 import { Batcher } from "./batcher";
 import { ExecutionEngine } from "./execution/execution-engine";
@@ -29,6 +28,7 @@ import { Reconciler } from "./reconciliation/reconciler";
 import { ArtifactMap } from "./reconciliation/types";
 import { isContractExecutionStateArray } from "./type-guards";
 import { ChainDispatcher } from "./types/chain-dispatcher";
+import { DeploymentLoader } from "./types/deployment-loader";
 import { ExecutionStrategy } from "./types/execution-engine";
 import {
   ContractAtExecutionState,
