+++ packages/core/src/new-api/internal/deployment-loader/file-deployment-loader.ts
@@ -2,8 +2,8 @@ import { ensureDir, pathExists, readFile, writeFile } from "fs-extra";
 import path from "path";
 
 import { Artifact, BuildInfo } from "../../types/artifact";
-import { DeploymentLoader } from "../../types/deployment-loader";
 import { FileJournal } from "../journal/file-journal";
+import { DeploymentLoader } from "../types/deployment-loader";
 import { Journal } from "../types/journal";
 
 export class FileDeploymentLoader implements DeploymentLoader {
