+++ packages/core/test/new-api/helpers.ts
@@ -7,6 +7,7 @@ import { Deployer } from "../../src/new-api/internal/deployer";
 import { AccountsState } from "../../src/new-api/internal/execution/execution-engine";
 import { MemoryJournal } from "../../src/new-api/internal/journal/memory-journal";
 import { ChainDispatcher } from "../../src/new-api/internal/types/chain-dispatcher";
+import { DeploymentLoader } from "../../src/new-api/internal/types/deployment-loader";
 import {
   OnchainState,
   OnchainStatuses,
@@ -20,7 +21,6 @@ import {
   DeploymentResultContract,
   DeploymentResultContracts,
 } from "../../src/new-api/types/deployer";
-import { DeploymentLoader } from "../../src/new-api/types/deployment-loader";
 
 export const exampleAccounts: string[] = [
   "0x70997970C51812dc3A010C7d01b50e0d17dc79C8",
