+++ packages/core/test/new-api/types/deployment-loader.ts
@@ -3,7 +3,7 @@ import { assert } from "chai";
 
 import { EphemeralDeploymentLoader } from "../../../src/new-api/internal/deployment-loader/ephemeral-deployment-loader";
 import { FileDeploymentLoader } from "../../../src/new-api/internal/deployment-loader/file-deployment-loader";
-import { DeploymentLoader } from "../../../src/new-api/types/deployment-loader";
+import { DeploymentLoader } from "../../../src/new-api/internal/types/deployment-loader";
 import { ExactInterface } from "../../helpers/exact-interface";
 import { setupMockArtifactResolver } from "../helpers";
